requirements.txt
amqp==5.3.1
aniso8601==7.0.0
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.1
django-bootstrap5==25.1
django-environ==0.12.0
django-filter==25.1
django-graphql-jwt==0.4.0
django-tables2==2.7.5
django_celery_results==2.6.0
graphene==3.4.3
graphene-django==3.2.2
graphql-core==3.2.6
graphql-relay==3.2.0
idna==3.8
kombu==5.5.2
promise==2.3
prompt_toolkit==3.0.50
pyee==12.1.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-ffmpeg==2.0.12
requests==2.32.3
Rx==1.6.3
singledispatch==4.1.1
six==1.17.0
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2025.2
Unidecode==1.3.8
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
